Fixed error when reading files
