+++ atmospheric_explorer/plotting/plot_utils.py
@@ -13,7 +13,6 @@ import plotly.graph_objects as go
 import xarray as xr
 
 from atmospheric_explorer.loggers import get_logger
-from atmospheric_explorer.utils import hex_to_rgb
 
 logger = get_logger("atmexp")
 
@@ -24,6 +23,24 @@ def _base_height(n_plots):
 
 def _row_spacing(n_plots):
     return 0.25 if n_plots > 3 else 0.3
+
+
+def hex_to_rgb(hex_color: str) -> tuple:
+    """\
+    Converts an hex color to a rgb tuple.
+    If an rgba color is provided, this function will return its rgb tuple and ignore the alpha channel.
+    """
+    if hex_color.startswith("#"):
+        hex_color = hex_color.lstrip("#")
+        if len(hex_color) == 3:
+            hex_color = hex_color * 2
+        return int(hex_color[0:2], 16), int(hex_color[2:4], 16), int(hex_color[4:6], 16)
+    return tuple(hex_color.strip("rgba()").split(",")[:3])
+
+
+def save_plotly_to_image(fig: go.Figure, path: str, img_format: str = "png") -> None:
+    """Save plotly plot to static image"""
+    fig.to_image(path, format=img_format)
 
 
 def _add_ci(
