+++ atmospheric_explorer/plotting/yearly_flux.py
@@ -9,13 +9,13 @@ import geopandas as gpd
 import plotly.graph_objects as go
 import xarray as xr
 
+from atmospheric_explorer.data_interface.data_transformations import (
+    clip_and_concat_shapes,
+    confidence_interval,
+)
 from atmospheric_explorer.data_interface.ghg import (
     GHGConfig,
     InversionOptimisedGreenhouseGas,
-)
-from atmospheric_explorer.data_transformations import (
-    clip_and_concat_shapes,
-    confidence_interval,
 )
 from atmospheric_explorer.loggers import get_logger
 from atmospheric_explorer.plotting.plot_utils import line_with_ci_subplots
