+++ tests/data_interface/test_cams_interfaces.py
@@ -13,7 +13,7 @@ def test__init():
     assert obj._file_format is None
     assert obj._file_ext is None
     assert obj._dataset_name is None
-    assert obj._id > 0
+    assert obj._id == 0
 
 
 def test_data_variables():
